app/requirement.txt
SQLAlchemy==2.0.15
requests-oauthlib==1.1.0
requests==2.31.0
Werkzeug==2.3.4
Flask==2.3.2
Flask-JWT==0.3.2
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.0.3
matplotlib==3.7.1
numpy==1.24.3
oauthlib==2.1.0
pip==23.1.2
plotly==5.14.1
PyGithub==1.58.2
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.1.0